Move tracker statistics from content blocker to site rating
