+++ DuckDuckGoTests/ContentBlockerTests.swift
@@ -28,8 +28,6 @@ class ContentBlockerTests: XCTestCase {
 
     struct UrlString {
         static let tracker = "http://www.atracker.com"
-        static let trackerWithPath = "http://www.atracker.com/aPath"
-        static let trackerWithAnotherPath = "http://www.atracker.com/anotherPath"
         static let differentTracker = "http://www.anothertracker.com"
         static let standard =  "http://www.facebook.com/path"
         static let document =  "http://www.wordpress.com"
@@ -37,8 +35,6 @@ class ContentBlockerTests: XCTestCase {
     
     struct Url {
         static let tracker = URL(string: UrlString.tracker)!
-        static let trackerWithPath = URL(string: UrlString.trackerWithPath)!
-        static let trackerWithAntotherPath = URL(string: UrlString.trackerWithAnotherPath)!
         static let differentTracker = URL(string: UrlString.differentTracker)!
         static let standard = URL(string: UrlString.standard)!
         static let document = URL(string: UrlString.document)!
@@ -56,14 +52,7 @@ class ContentBlockerTests: XCTestCase {
         mockConifguartion.trackers = trackers()
         testee = ContentBlocker(configuration: mockConifguartion)
     }
-
-    func testCountsAreInitiallyZero() {
-        XCTAssertEqual(testee.totalItemsDetected, 0)
-        XCTAssertEqual(testee.uniqueItemsBlocked, 0)
-        XCTAssertEqual(testee.totalItemsBlocked, 0)
-        XCTAssertEqual(testee.uniqueItemsBlocked, 0)
-    }
-    
+  
     func testWhenEnabledThenEnabledIsTrue() {
         testee.enabled = true
         XCTAssertTrue(testee.enabled)
@@ -83,18 +72,24 @@ class ContentBlockerTests: XCTestCase {
         XCTAssertTrue(testee.hasData)
     }
     
-    func testWhenTrackerUrlThenBlocked() {
-        XCTAssertTrue(testee.block(url: Url.tracker, forDocument: Url.document))
+    func testWhenTrackerUrlThenTrackerBlocked() {
+        let policy = testee.policy(forUrl: Url.tracker, document: Url.document)
+        XCTAssertNotNil(policy.tracker)
+        XCTAssertTrue(policy.block)
     }
     
-    func testWhenBlockerDisabledThenNotBlocked() {
+    func testWhenBlockerDisabledThenTrackerNotBlocked() {
         mockConifguartion.enabled = false
-        XCTAssertFalse(testee.block(url: Url.tracker, forDocument: Url.document))
+        let policy = testee.policy(forUrl: Url.tracker, document: Url.document)
+        XCTAssertNotNil(policy.tracker)
+        XCTAssertFalse(policy.block)
     }
 
     func testWhenDomainAddedToWhitelistThenNoTrackersBlocked() {
         testee.whitelist(true, domain: Url.document.host!)
-        XCTAssertFalse(testee.block(url: Url.tracker, forDocument: Url.document))
+        let policy = testee.policy(forUrl: Url.tracker, document: Url.document)
+        XCTAssertNotNil(policy.tracker)
+        XCTAssertFalse(policy.block)
     }
     
     func testWhenDomainWhitelistedThenEnabledForDomainIsFalse() {
@@ -114,59 +109,21 @@ class ContentBlockerTests: XCTestCase {
         let domain = Url.document.host!
         testee.whitelist(true, domain: domain)
         testee.whitelist(false, domain: Url.document.host!)
-        XCTAssertTrue(testee.block(url: Url.tracker, forDocument: Url.document))
+        
+        let policy = testee.policy(forUrl: Url.tracker, document: Url.document)
+        XCTAssertNotNil(policy.tracker)
+        XCTAssertTrue(policy.block)
     }
     
     func testWhenFirstPartyUrlThenNotBlocked() {
-        XCTAssertFalse(testee.block(url: Url.tracker, forDocument: Url.tracker))
+        let policy = testee.policy(forUrl: Url.tracker, document: Url.tracker)
+        XCTAssertNil(policy.tracker)
+        XCTAssertFalse(policy.block)
     }
     
     func testWhenNonTrackerUrlThenNotBlocked() {
-        XCTAssertFalse(testee.block(url: Url.standard, forDocument: Url.document))
-    }
-
-    func testWhenUniqueTrackersAreBlockedThenAllDetectionAndBlockCountsIncremenet() {
-        XCTAssertTrue(testee.block(url: Url.tracker, forDocument: Url.document))
-        XCTAssertTrue(testee.block(url: Url.differentTracker, forDocument: Url.document))
-        XCTAssertEqual(testee.totalItemsDetected, 2)
-        XCTAssertEqual(testee.uniqueItemsDetected, 2)
-        XCTAssertEqual(testee.totalItemsBlocked, 2)
-        XCTAssertEqual(testee.uniqueItemsBlocked, 2)
-    }
-
-    func testWhenRepeatTrackersAreBlockedThenUniqueCountsOnlyIncrementOnce() {
-        XCTAssertTrue(testee.block(url: Url.tracker, forDocument: Url.document))
-        XCTAssertTrue(testee.block(url: Url.tracker, forDocument: Url.document))
-        XCTAssertEqual(testee.totalItemsDetected, 2)
-        XCTAssertEqual(testee.uniqueItemsDetected, 1)
-        XCTAssertEqual(testee.totalItemsBlocked, 2)
-        XCTAssertEqual(testee.uniqueItemsBlocked, 1)
-    }
-    
-    func testWhenTrackersWithSameDomainAreBlockedThenUniqueCountsOnlyIncrementOnce() {
-        XCTAssertTrue(testee.block(url: Url.trackerWithPath, forDocument: Url.document))
-        XCTAssertTrue(testee.block(url: Url.trackerWithAntotherPath, forDocument: Url.document))
-        XCTAssertEqual(testee.totalItemsDetected, 2)
-        XCTAssertEqual(testee.uniqueItemsDetected, 1)
-        XCTAssertEqual(testee.totalItemsBlocked, 2)
-        XCTAssertEqual(testee.uniqueItemsBlocked, 1)
-    }
-    
-    func testWhenBlockerDisabledThenDetectedCountsIncrementButBlockCountsDoNot() {
-        mockConifguartion.enabled = false
-        XCTAssertFalse(testee.block(url: Url.tracker, forDocument: Url.document))
-        XCTAssertEqual(testee.totalItemsDetected, 1)
-        XCTAssertEqual(testee.uniqueItemsDetected, 1)
-        XCTAssertEqual(testee.totalItemsBlocked, 0)
-        XCTAssertEqual(testee.uniqueItemsBlocked, 0)
-    }
-    
-    func testResetClearsCounts() {
-        testWhenUniqueTrackersAreBlockedThenAllDetectionAndBlockCountsIncremenet()
-        testee.resetMonitoring()
-        XCTAssertEqual(testee.totalItemsDetected, 0)
-        XCTAssertEqual(testee.uniqueItemsDetected, 0)
-        XCTAssertEqual(testee.totalItemsBlocked, 0)
-        XCTAssertEqual(testee.uniqueItemsBlocked, 0)
+        let policy = testee.policy(forUrl: Url.standard, document: Url.document)
+        XCTAssertNil(policy.tracker)
+        XCTAssertFalse(policy.block)
     }
 }
